requirements.txt
attrs==20.3.0
docutils==0.17
iniconfig==1.1.1
lockfile==0.12.2
packaging==20.9
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-nmap==0.6.4
toml==0.10.2
